refactor: generator를 model로 이동
