+++ src/main/java/lotto/model/generator/RandomLottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.util;
+package lotto.model.generator;
 
 import java.util.ArrayList;
 import java.util.Collections;
